+++ CHANGES.rst
@@ -61,6 +61,16 @@ Changelog
 - Added possibility to make a committee selectable only on an item and
   not on a meeting.
   [gbastien]
+- Added adaptable method `MeetingItem._annex_decision_addable_states_after_validation`
+  that will manage item states in which annex decision may be added after the
+  validation process so since the `validated` state until the end of the item WF.
+  [gbastien]
+- Added WF adaptation `waiting_advices_given_and_signed_advices_required_to_validate`
+  that will check if necessary advice reached their WF last step.
+  This is an answer to rare case where advice is not given `completely` and item was
+  validated, now if advice WF last step was no reached, it will not be possible to
+  validate the item.
+  [gbastien]
 - On the meeting view, when no available items, close the `available-items`
   collapsible so it takes less place and display the number of available items
   like it is already the case for presented items so it is clear why the